Refactoring of qualification test __init__ and conftest
They were getting a bit unwieldy, so I moved CBFRemoteControl and CBFCache
(and their dependencies) out to a separate cbf.py. That leaves conftest
just with actual fixtures rather than helper classes and functions.

For now __init__ reimports CBFRemoteControl so that the rest of the
tests don't need to change their import line. I'll update that shortly.
